requirements.txt
# FastAPI Backend
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & ML
numpy>=1.26.0
scipy>=1.11.4
scikit-learn>=1.3.2
# torch>=2.1.0  # Uncomment if you need PyTorch (large download)
# torchvision>=0.16.0
# tensorflow>=2.15.0  # Uncomment if you need TensorFlow (large download)

# GIS & Point Cloud Processing
# gdal>=3.7.3  # Requires special installation on Windows - see INSTALL.md
rasterio>=1.3.9
laspy>=2.5.1
# pdal>=3.7.0  # Requires special installation
shapely>=2.0.2
geopandas>=0.14.1
# fiona>=1.9.5  # Requires GDAL

# Image Processing
opencv-python>=4.8.1
Pillow>=10.1.0

# DEM & Contour Generation
pyproj>=3.6.1
matplotlib>=3.8.2
scikit-image>=0.22.0

# Export Formats
ezdxf>=1.1.0
pyshp>=2.3.1

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.2.1
